Format javascript and solidity files

+++ contracts/BN254.sol
@@ -11,26 +11,29 @@ pragma solidity ^0.8.0;
 
 library Curve {
     // p = p(u) = 36u^4 + 36u^3 + 24u^2 + 6u + 1
-    uint256 internal constant FIELD_ORDER = 0x30644e72e131a029b85045b68181585d97816a916871ca8d3c208c16d87cfd47;
+    uint256 internal constant FIELD_ORDER =
+        0x30644e72e131a029b85045b68181585d97816a916871ca8d3c208c16d87cfd47;
 
     // Number of elements in the field (often called `q`)
     // n = n(u) = 36u^4 + 36u^3 + 18u^2 + 6u + 1
-    uint256 internal constant GEN_ORDER = 0x30644e72e131a029b85045b68181585d2833e84879b9709143e1f593f0000001;
+    uint256 internal constant GEN_ORDER =
+        0x30644e72e131a029b85045b68181585d2833e84879b9709143e1f593f0000001;
 
     uint256 internal constant CURVE_B = 3;
 
     // a = (p+1) / 4
-    uint256 internal constant CURVE_A = 0xc19139cb84c680a6e14116da060561765e05aa45a1c72a34f082305b61f3f52;
+    uint256 internal constant CURVE_A =
+        0xc19139cb84c680a6e14116da060561765e05aa45a1c72a34f082305b61f3f52;
 
     struct G1Point {
-        uint X;
-        uint Y;
+        uint256 X;
+        uint256 Y;
     }
 
     // Encoding of field elements is: X[0] * z + X[1]
     struct G2Point {
-        uint[2] X;
-        uint[2] Y;
+        uint256[2] X;
+        uint256[2] Y;
     }
 
     // (P+1) / 4
@@ -51,20 +54,18 @@ library Curve {
         return G1Point(1, 2);
     }
 
-    function HashToPoint(uint256 s)
-    internal returns (G1Point memory)
-    {
+    function HashToPoint(uint256 s) internal returns (G1Point memory) {
         uint256 beta = 0;
         uint256 y = 0;
 
         // XXX: Gen Order (n) or Field Order (p) ?
         uint256 x = s % GEN_ORDER;
 
-        while( true ) {
+        while (true) {
             (beta, y) = FindYforX(x);
 
             // y^2 == beta
-            if( beta == mulmod(y, y, FIELD_ORDER) ) {
+            if (beta == mulmod(y, y, FIELD_ORDER)) {
                 return G1Point(x, y);
             }
 
@@ -72,19 +73,20 @@ library Curve {
         }
     }
 
-
     /**
-    * Given X, find Y
-    *
-    *   where y = sqrt(x^3 + b)
-    *
-    * Returns: (x^3 + b), y
-    */
-    function FindYforX(uint256 x)
-    internal returns (uint256, uint256)
-    {
+     * Given X, find Y
+     *
+     *   where y = sqrt(x^3 + b)
+     *
+     * Returns: (x^3 + b), y
+     */
+    function FindYforX(uint256 x) internal returns (uint256, uint256) {
         // beta = (x^3 + b) % p
-        uint256 beta = addmod(mulmod(mulmod(x, x, FIELD_ORDER), x, FIELD_ORDER), CURVE_B, FIELD_ORDER);
+        uint256 beta = addmod(
+            mulmod(mulmod(x, x, FIELD_ORDER), x, FIELD_ORDER),
+            CURVE_B,
+            FIELD_ORDER
+        );
 
         // y^2 = x^3 + b
         // this acts like: y = sqrt(beta)
@@ -93,59 +95,72 @@ library Curve {
         return (beta, y);
     }
 
-
     // a - b = c;
-    function submod(uint a, uint b) internal returns (uint){
-        uint a_nn;
+    function submod(uint256 a, uint256 b) internal returns (uint256) {
+        uint256 a_nn;
 
-        if(a>b) {
+        if (a > b) {
             a_nn = a;
         } else {
-            a_nn = a+GEN_ORDER;
+            a_nn = a + GEN_ORDER;
         }
 
         return addmod(a_nn - b, 0, GEN_ORDER);
     }
 
-
-    function expMod(uint256 _base, uint256 _exponent, uint256 _modulus)
-    internal returns (uint256 retval)
-    {
+    function expMod(
+        uint256 _base,
+        uint256 _exponent,
+        uint256 _modulus
+    ) internal returns (uint256 retval) {
         bool success;
         uint256[1] memory output;
-        uint[6] memory input;
-        input[0] = 0x20;        // baseLen = new(big.Int).SetBytes(getData(input, 0, 32))
-        input[1] = 0x20;        // expLen  = new(big.Int).SetBytes(getData(input, 32, 32))
-        input[2] = 0x20;        // modLen  = new(big.Int).SetBytes(getData(input, 64, 32))
+        uint256[6] memory input;
+        input[0] = 0x20; // baseLen = new(big.Int).SetBytes(getData(input, 0, 32))
+        input[1] = 0x20; // expLen  = new(big.Int).SetBytes(getData(input, 32, 32))
+        input[2] = 0x20; // modLen  = new(big.Int).SetBytes(getData(input, 64, 32))
         input[3] = _base;
         input[4] = _exponent;
         input[5] = _modulus;
         assembly {
-            success := staticcall(sub(gas(), 2000), 5, input, 0xc0, output, 0x20)
+            success := staticcall(
+                sub(gas(), 2000),
+                5,
+                input,
+                0xc0,
+                output,
+                0x20
+            )
             // Use "invalid" to make gas estimation work
-            switch success case 0 { revert(0,0) }
+            switch success
+            case 0 {
+                revert(0, 0)
+            }
         }
         require(success);
         return output[0];
     }
 
-
     /// @return the generator of G2
     function P2() internal returns (G2Point memory) {
-        return G2Point(
-            [11559732032986387107991004021392285783925812861821192530917403151452391805634,
-            10857046999023057135944570762232829481370756359578518086990519993285655852781],
-            [4082367875863433681332203403145435568316851327593401208105741076214120093531,
-            8495653923123431417604973247489272438418190587263600148770280649306958101930]
-        );
+        return
+            G2Point(
+                [
+                    11559732032986387107991004021392285783925812861821192530917403151452391805634,
+                    10857046999023057135944570762232829481370756359578518086990519993285655852781
+                ],
+                [
+                    4082367875863433681332203403145435568316851327593401208105741076214120093531,
+                    8495653923123431417604973247489272438418190587263600148770280649306958101930
+                ]
+            );
     }
 
     /// @return the negation of p, i.e. p.add(p.negate()) should be zero.
-    function g1neg(G1Point memory p)  internal returns (G1Point memory) {
+    function g1neg(G1Point memory p) internal returns (G1Point memory) {
         // The prime q in the base field F_q for G1
-        uint q = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
-        if (p.X == 0 && p.Y == 0)
-            return G1Point(0, 0);
+        uint256 q = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
+        if (p.X == 0 && p.Y == 0) return G1Point(0, 0);
         return G1Point(p.X, q - (p.Y % q));
     }
 
@@ -154,8 +169,11 @@ library Curve {
      * @param p2 G1 point
      * @returns the sum of two points of G1
      */
-    function g1add(G1Point memory p1, G1Point memory p2) internal returns (G1Point memory r) {
-        uint[4] memory input;
+    function g1add(G1Point memory p1, G1Point memory p2)
+        internal
+        returns (G1Point memory r)
+    {
+        uint256[4] memory input;
         input[0] = p1.X;
         input[1] = p1.Y;
         input[2] = p2.X;
@@ -163,8 +181,11 @@ library Curve {
         bool success;
         assembly {
             success := staticcall(sub(gas(), 2000), 6, input, 0xc0, r, 0x60)
-        // Use "invalid" to make gas estimation work
-            switch success case 0 { revert(0,0) }
+            // Use "invalid" to make gas estimation work
+            switch success
+            case 0 {
+                revert(0, 0)
+            }
         }
         require(success);
     }
@@ -175,31 +196,39 @@ library Curve {
      * @returns the product of a point on G1 and a scalar, i.e.
      */
     /// p == p.mul(1) and p.add(p) == p.mul(2) for all points p.
-    function g1mul(G1Point memory p, uint s)  internal returns (G1Point memory r) {
-        uint[3] memory input;
+    function g1mul(G1Point memory p, uint256 s)
+        internal
+        returns (G1Point memory r)
+    {
+        uint256[3] memory input;
         input[0] = p.X;
         input[1] = p.Y;
         input[2] = s;
         bool success;
         assembly {
             success := staticcall(sub(gas(), 2000), 7, input, 0x80, r, 0x60)
-        // Use "invalid" to make gas estimation work
-            switch success case 0 { revert(0,0) }
+            // Use "invalid" to make gas estimation work
+            switch success
+            case 0 {
+                revert(0, 0)
+            }
         }
-        require (success);
+        require(success);
     }
 
     /// @return the result of computing the pairing check
     /// e(p1[0], p2[0]) *  .... * e(p1[n], p2[n]) == 1
     /// For example pairing([P1(), P1().negate()], [P2(), P2()]) should
     /// return true.
-    function pairing(G1Point[] memory p1, G2Point[] memory p2) internal returns (bool) {
+    function pairing(G1Point[] memory p1, G2Point[] memory p2)
+        internal
+        returns (bool)
+    {
         require(p1.length == p2.length);
-        uint elements = p1.length;
-        uint inputSize = elements * 6;
-        uint[] memory input = new uint[](inputSize);
-        for (uint i = 0; i < elements; i++)
-        {
+        uint256 elements = p1.length;
+        uint256 inputSize = elements * 6;
+        uint256[] memory input = new uint256[](inputSize);
+        for (uint256 i = 0; i < elements; i++) {
             input[i * 6 + 0] = p1[i].X;
             input[i * 6 + 1] = p1[i].Y;
             input[i * 6 + 2] = p2[i].X[0];
@@ -207,12 +236,22 @@ library Curve {
             input[i * 6 + 4] = p2[i].Y[0];
             input[i * 6 + 5] = p2[i].Y[1];
         }
-        uint[1] memory out;
+        uint256[1] memory out;
         bool success;
         assembly {
-            success := staticcall(sub(gas(), 2000), 8, add(input, 0x20), mul(inputSize, 0x20), out, 0x20)
-        // Use "invalid" to make gas estimation work
-            switch success case 0 { revert(0,0) }
+            success := staticcall(
+                sub(gas(), 2000),
+                8,
+                add(input, 0x20),
+                mul(inputSize, 0x20),
+                out,
+                0x20
+            )
+            // Use "invalid" to make gas estimation work
+            switch success
+            case 0 {
+                revert(0, 0)
+            }
         }
         require(success);
         return out[0] != 0;
